LPS-57770 Move
