+++ modules/portal-cache/portal-cache-cluster/src/com/liferay/portal/cache/cluster/internal/ClusterLinkPortalCacheReplicator.java
@@ -14,7 +14,7 @@
 
 package com.liferay.portal.cache.cluster.internal;
 
-import com.liferay.portal.cache.cluster.internal.clusterlink.PortalCacheClusterLink;
+import com.liferay.portal.cache.cluster.internal.cluster.link.PortalCacheClusterLink;
 import com.liferay.portal.kernel.cache.PortalCache;
 import com.liferay.portal.kernel.cache.PortalCacheException;
 import com.liferay.portal.kernel.cache.PortalCacheManager;
