+++ modules/portal-cache/portal-cache-cluster/src/com/liferay/portal/cache/cluster/internal/ClusterLinkPortalCacheReplicatorFactory.java
@@ -14,7 +14,7 @@
 
 package com.liferay.portal.cache.cluster.internal;
 
-import com.liferay.portal.cache.cluster.internal.clusterlink.PortalCacheClusterLink;
+import com.liferay.portal.cache.cluster.internal.cluster.link.PortalCacheClusterLink;
 import com.liferay.portal.kernel.cache.PortalCacheReplicator;
 import com.liferay.portal.kernel.cache.PortalCacheReplicatorFactory;
