+++ modules/portal-cache/portal-cache-cluster/src/com/liferay/portal/cache/cluster/internal/cluster/link/ClusterLinkPortalCacheClusterChannelFactory.java
@@ -12,18 +12,37 @@
  * details.
  */
 
-package com.liferay.portal.cache.cluster.internal.clusterlink;
+package com.liferay.portal.cache.cluster.internal.cluster.link;
 
+import com.liferay.portal.cache.cluster.internal.DestinationNames;
 import com.liferay.portal.cache.cluster.internal.PortalCacheClusterException;
+import com.liferay.portal.kernel.cluster.ClusterLink;
 import com.liferay.portal.kernel.cluster.Priority;
+
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
 
 /**
  * @author Shuyang Zhou
  */
-public interface PortalCacheClusterChannelFactory {
+@Component(immediate = true, service = PortalCacheClusterChannelFactory.class)
+public class ClusterLinkPortalCacheClusterChannelFactory
+	implements PortalCacheClusterChannelFactory {
 
+	@Override
 	public PortalCacheClusterChannel createPortalCacheClusterChannel(
 			Priority priority)
-		throws PortalCacheClusterException;
+		throws PortalCacheClusterException {
+
+		return new ClusterLinkPortalCacheClusterChannel(
+			_clusterLink, DestinationNames.CACHE_REPLICATION, priority);
+	}
+
+	@Reference(unbind = "-")
+	protected void setClusterLink(ClusterLink clusterLink) {
+		_clusterLink = clusterLink;
+	}
+
+	private ClusterLink _clusterLink;
 
 }
